refactor(constants): share environment short names
